pyproject.toml
[tool.poetry]
name = "tenants"
version = "0.1.0"
description = ""
authors = ["Vaibhav-Lodha <vaibhav.lodha@squadstack.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
django-tenants = "^3.6.1"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
